requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.23
django-debug-toolbar==4.4.6
django-modeltranslation==0.19.13
idna==3.10
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.13.2
urllib3==2.2.3
